Added Knockout-Validation plugin and configured validation on Fund entity.

+++ MvcWebRole/App/viewmodels/funds/create.js
@@ -25,9 +25,11 @@
         //#region Internal methods.
         function activate(queryString) {
             logger.log('Create fund view activated', null, 'funds/create', false);
-            vm.item(datacontext.createItem({
+            vm.item = datacontext.createItem({
                 AreaId: queryString.areaid
-            }));
+            });
+
+            vm.errors = ko.validation.group(vm.item());
             return true;
         }
 
@@ -52,24 +54,28 @@
 
         // TODO: Client-side validation.
         function saveItem(item) {
-            // Remove uploads with errors.
-            var uploadItemsWithErrors = item.FileUploads.remove(function (uploadItem) {
-                return uploadItem.errorMessage();
-            });
+            if (vm.errors().length == 0) {
+                // Remove uploads with errors.
+                var uploadItemsWithErrors = item.FileUploads.remove(function (uploadItem) {
+                    return uploadItem.errorMessage();
+                });
 
-            // Remove uploads marked with destroy.
-            var removedUploadItems = item.FileUploads.remove(function (uploadItem) {
-                return uploadItem.destroy();
-            });
+                // Remove uploads marked with destroy.
+                var removedUploadItems = item.FileUploads.remove(function (uploadItem) {
+                    return uploadItem.destroy();
+                });
 
-            // Delete removed files from server.
-            $.each(removedUploadItems, function (index, value) {
-                fileuploadDatacontext.deleteItem(value);
-            });
+                // Delete removed files from server.
+                $.each(removedUploadItems, function (index, value) {
+                    fileuploadDatacontext.deleteItem(value);
+                });
 
-            datacontext.saveNewItem(
-                item,
-                [addNewItemToBrowseVM, navigateToBrowseView]);
+                datacontext.saveNewItem(
+                    item,
+                    [addNewItemToBrowseVM, navigateToBrowseView]);
+            } else {
+                vm.errors.showAllMessages();
+            }
         }
 
         function addNewItemToBrowseVM(newItem) {
